A lot of changes

+++ database/migrations/2025_05_11_114840_add_recruited_to_post_user_applications.php
@@ -1,26 +1,23 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
-    public function up()
-    {
+return new class extends Migration {
+    public function up(): void {
+        if (!Schema::hasTable('post_user_applications')) return;
         Schema::table('post_user_applications', function (Blueprint $table) {
-            $table->boolean('recruited')->default(false);
+            if (!Schema::hasColumn('post_user_applications', 'recruited')) {
+                $table->boolean('recruited')->default(false);
+            }
         });
     }
-
-    public function down()
-    {
+    public function down(): void {
+        if (!Schema::hasTable('post_user_applications')) return;
         Schema::table('post_user_applications', function (Blueprint $table) {
-            $table->dropColumn('recruited');
+            if (Schema::hasColumn('post_user_applications', 'recruited')) {
+                $table->dropColumn('recruited');
+            }
         });
     }
-
 };
